Replace the backward-kill-word command
cdip-backward-kill-word:

If the region is active, then invoke kill-region. Otherwise, use the
following custom backward-kill-word procedure.  If the previous word is
on the same line, then kill the previous word.  Otherwise, if the
previous word is on a prior line, then kill to the beginning of the
line.  If point is already at the beginning of the line, then kill to
the end of the previous line.

With argument ARG and region inactive, do this that many times.